berliner_frauen/pyproject.toml
[tool.poetry]
name = "berliner_frauen"
version = "0.1.0"
description = ""
authors = ["Rosamund Mather <rosamund@posteo.de>"]

[tool.poetry.dependencies]
aniso8601 = "^7.0.0"
appdirs = "^1.4.4"
asgiref = "^3.4.1"
attrs = "^21.4.0"
beautifulsoup4 = "^4.10.0"
bs4 = "^0.0.1"
CacheControl = "^0.12.10"
cachy = "^0.3.0"
certifi = "^2021.10.8"
charset-normalizer = "^2.0.7"
cleo = "^0.8.1"
click = "^7.1.2"
clikit = "^0.6.2"
crashtest = "^0.3.1"
distlib = "^0.3.3"
Django = "^3.2.9"
django-cors-headers = "^3.10.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
filelock = "^3.3.2"
html5lib = "^1.1"
idna = "^3.3"
keyring = "^21.8.0"
lockfile = "^0.12.2"
more-itertools = "^8.12.0"
msgpack = "^1.0.2"
mypy-extensions = "^0.4.3"
packaging = "^20.9"
pastel = "^0.2.1"
pathspec = "^0.8.1"
pexpect = "^4.8.0"
pkginfo = "^1.7.1"
platformdirs = "^2.4.0"
pluggy = "^0.13.1"
poetry = "^1.1.12"
poetry-core = "^1.0.7"
promise = "^2.3"
ptyprocess = "^0.7.0"
py = "^1.11.0"
PyJWT = "^2.3.0"
pylev = "^1.4.0"
pyparsing = "^3.0.7"
python-dateutil = "^2.8.2"
python = "^3.7"
pytz = "^2021.3"
regex = "^2020.11.13"
requests = "^2.26.0"
requests-toolbelt = "^0.9.1"
Rx = "^1.6.1"
shellingham = "^1.4.0"
singledispatch = "^3.7.0"
six = "^1.16.0"
soupsieve = "^2.3.1"
sqlparse = "^0.4.2"
text-unidecode = "^1.3"
toml = "^0.10.2"
tomlkit = "^0.7.2"
typed-ast = "^1.4.1"
typing-extensions = "^3.7.4.3"
urllib3 = "^1.26.7"
virtualenv = "^20.10.0"
wcwidth = "^0.2.5"
webencodings = "^0.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
